fix: `link-app` script breaking consumers tests  (#154)
* Manually create styled-components link

* Use `yarn link` to link styled-components

* Remove duplicate packages from root package.json

+++ tasks/unlink-app.sh
@@ -1,17 +1,31 @@
 #!/usr/bin/env bash
 
+if [ -z "$1" ]; then
+  echo "Usage: $0 [path-to-consuming-app]"
+  exit 1
+fi
+
+roo_path="$(cd `dirname $0`/.. && pwd)"
+consumer_path="$(cd $1 && pwd)"
 packages=($(yarn lerna ls | grep @roo-ui))
-packages+=('styled-components')
 
-cd $1
+cd "$consumer_path"
 
+# Remove links in comsumer
+cd $consumer_path
 for package in ${packages[@]/*v*/}; do
   yarn unlink "$package"
 done
 
+yarn unlink "styled-components"
+
 echo "reinstall dependencies"
 yarn --check-files
 
-cd -
+cd "$roo_path"
 
 yarn lerna exec 'yarn unlink'
+
+cd "$roo_path/node_modules/styled-components"
+
+yarn unlink
